alloc_tracker: estimate the peak the same way as the other metrics
Taking the maximum span watermark made the peak maximally sensitive to
exactly the spans that deserve the least trust. A benchmark harness runs
short warmup batches before settling into long steady-state ones, and a
one-iteration warmup batch with an anomalous watermark won the maximum
outright.

Estimate it instead through the shared span accumulator, which weights
each span by the square of its iteration count and so all but discards
short batches without needing to know which ones were warmup. A
watermark does not scale with the iteration count, so it is scaled up by
that count on the way in and divided back out by the regression, leaving
the span watermarks averaged with weight n².

The peak consequently becomes a per-iteration figure like its
neighbours, carries a confidence interval, and joins the JSON output in
the same slope-and-interval shape the other metrics use. In exchange it
now assumes every iteration within a batch reaches the same peak, which
is documented.

Copilot-Session: f9527c14-db4b-410b-b77f-16c19b6ecf29

+++ packages/alloc_tracker/src/operation.rs
@@ -171,10 +171,10 @@ impl Operation {
         data.total_bytes_allocated()
     }
 
-    /// Returns the most bytes held allocated at once across all recorded spans.
+    /// Returns the per-iteration peak bytes held allocated across all recorded spans.
     #[must_use]
     #[cfg(test)]
-    pub(crate) fn peak_outstanding_bytes(&self) -> Option<u64> {
+    pub(crate) fn peak_outstanding_bytes(&self) -> Option<f64> {
         let data = self.metrics.lock().expect(ERR_POISONED_LOCK);
         data.peak_outstanding_bytes()
     }
